Add version using gmp internally (but in-out is double).
Forwarding gmp values to python side is difficult, so try this compromise.
Hopefully it can be sufficient.

Had to fix a type error in Eigen ; gmp generates template expressions,
which did not play well with some type constraints on an internal
function.
Added a forced type conversion to fix it.