Fix pybind11 C-API breakage with python3.11
Python 3.11 breaks the C-API.
pybind11 2.10 fixes the breakage.

To prevent future problems of this type, pybind11 is now a build
requirement instead of a hardcopy included in git.
To do that, the project has been converted to the new python build
standard (PEP517, pyproject.toml).

This fix is not perfect as paths may be ambiguous if a system version of
pybind11 is installed ; better to remove it if unneeded.