Fixed broken metadata
I tried to be clever with url metadata formatting but this did not work.
Revert to official doc formatting.

Eigen situation :
2 packages exist that provide eigen header files.
Both only have one release, one is from 2021 (eigenpip) and the other is 2023 (includeigen).
The 2021 one is probably unmaintained.
Hard to say if the 2023 one will be maintained.

Thus for now:
- keep the included eigen version as it is working
- do not switch to any of these packages
- keep a note for the 2023 one in case we need it later